improvements v4
