+++ packages/scan/src/new-outlines/index.ts
@@ -4,27 +4,11 @@ import {
   getDisplayName,
   getFiberId,
   getNearestHostFibers,
-  getTimings,
-  getType,
   isCompositeFiber,
 } from 'bippy';
-import {
-  type Change,
-  type ContextChange,
-  type PropsChange,
-  ReactScanInternals,
-  Store,
-  ignoredProps,
-} from '~core/index';
-import {
-  ChangeReason,
-  createInstrumentation,
-  getContextChanges,
-  getStateChanges,
-} from '~core/instrumentation';
-import type { RenderData } from '~core/utils';
+import { ReactScanInternals, Store, ignoredProps } from '~core/index';
+import { createInstrumentation } from '~core/instrumentation';
 import { inspectorUpdateSignal } from '~web/components/inspector/states';
-import { getChangedPropsDetailed } from '~web/components/inspector/utils';
 import { readLocalStorage, removeLocalStorage } from '~web/utils/helpers';
 import { log, logIntro } from '~web/utils/log';
 import {
@@ -437,75 +421,6 @@ export const startReportInterval = () => {
   }, 50);
 };
 
-const reportRenderToListeners = (fiber: Fiber) => {
-  if (isCompositeFiber(fiber)) {
-    // report render has a non trivial cost because it calls Date.now(), so we want to avoid the computation if possible
-    if (
-      ReactScanInternals.options.value.showToolbar !== false &&
-      Store.inspectState.value.kind === 'focused'
-    ) {
-      const reportFiber = fiber;
-      const { selfTime } = getTimings(fiber);
-      const displayName = getDisplayName(fiber.type);
-      const fiberId = getFiberId(reportFiber);
-
-      const currentData = Store.reportData.get(fiberId);
-      const existingCount = currentData?.count ?? 0;
-      const existingTime = currentData?.time ?? 0;
-
-      const changes: Array<Change> = [];
-
-      // optimization, for now only track changes on inspected prop, cleanup later when changes is used in outline drawing
-      const listeners = Store.changesListeners.get(getFiberId(fiber));
-
-      if (listeners?.length) {
-        const propsChanges: Array<PropsChange> = getChangedPropsDetailed(fiber).map(
-          (change) => ({
-            type: ChangeReason.Props,
-            name: change.name,
-            value: change.value,
-            prevValue: change.prevValue,
-            unstable: false,
-          }),
-        );
-
-        const stateChanges = getStateChanges(fiber);
-
-        // context changes are incorrect, bippy needs to tell us the context dependencies that changed and provide those values every render
-        // currently, we say every context change, regardless of the render it happened, is a change. Which requires us to hack change tracking
-        // in the whats-changed toolbar component
-        const fiberContext = getContextChanges(fiber);
-        const contextChanges: Array<ContextChange> = fiberContext.map(
-          (info) => ({
-            name: info.name,
-            type: ChangeReason.Context,
-            value: info.value,
-            contextType: info.contextType,
-          }),
-        );
-
-        for (const listener of listeners) {
-          listener({
-            propsChanges,
-            stateChanges,
-            contextChanges,
-          });
-        }
-      }
-      const fiberData: RenderData = {
-        count: existingCount + 1,
-        time: existingTime + selfTime || 0,
-        renders: [],
-        displayName,
-        type: getType(fiber.type) || null,
-        changes,
-      };
-
-      Store.reportData.set(fiberId, fiberData);
-      needsReport = true;
-    }
-  }
-};
 export const isValidFiber = (fiber: Fiber) => {
   if (ignoredProps.has(fiber.memoizedProps)) {
     return false;
@@ -555,13 +470,11 @@ export const initReactScanInstrumentation = () => {
         // this can be expensive given enough re-renders
         log(renders);
       }
-      // optimization, only valid if toolbar is only listener
-      if (!isInspectorInactive) {
-        reportRenderToListeners(fiber);
-      }
 
       if (Store.inspectState.value.kind === 'focused') {
-        inspectorUpdateSignal.value = Date.now();
+        if (isCompositeFiber(fiber)) {
+          inspectorUpdateSignal.value = Date.now();
+        }
       }
 
       ReactScanInternals.options.value.onRender?.(fiber, renders);
