+++ packages/scan/src/web/components/inspector/utils.ts
@@ -1,5 +1,7 @@
 import {
   type Fiber,
+  FunctionComponentTag,
+  type MemoizedState,
   getDisplayName,
   getTimings,
   isCompositeFiber,
@@ -12,15 +14,9 @@ import { isEqual } from '~core/utils';
 import { batchGetBoundingRects } from '~web/utils/outline';
 import { globalInspectorState } from '.';
 import type { ExtendedReactRenderer } from '../../../types';
-import {
-  ensureRecord,
-  getAllFiberContexts,
-  getCurrentFiberState,
-  getStateNames,
-  isPromise,
-} from './overlay/utils';
 import { TIMELINE_MAX_UPDATES } from './states';
 import type { MinimalFiberInfo } from './states';
+import { getAllFiberContexts, getStateNames } from './timeline/utils';
 
 interface StateItem {
   name: string;
@@ -1534,6 +1530,276 @@ export const calculateSliderValues = (
     value: currentIndex,
     rightValue: totalUpdates - 1,
   };
+};
+
+// be careful, this is an implementation detail is not stable or reliable across all react versions https://github.com/facebook/react/pull/15124
+// type UpdateQueue<S, A> = {
+//   last: Update<S, A> | null,
+//   dispatch: (A => mixed) | null,
+//   eagerReducer: ((S, A) => S) | null,
+//   eagerState: S | null,
+// };
+
+interface ExtendedMemoizedState extends MemoizedState {
+  queue?: {
+    lastRenderedState: unknown;
+  } | null;
+  element?: unknown;
+}
+
+export const isDirectComponent = (fiber: Fiber): boolean => {
+  if (!fiber || !fiber.type) return false;
+
+  const isFunctionalComponent = typeof fiber.type === 'function';
+  const isClassComponent = fiber.type?.prototype?.isReactComponent ?? false;
+
+  if (!(isFunctionalComponent || isClassComponent)) return false;
+
+  if (isClassComponent) {
+    return true;
+  }
+
+  let memoizedState = fiber.memoizedState;
+  while (memoizedState) {
+    if (memoizedState.queue) {
+      return true;
+    }
+    const nextState: ExtendedMemoizedState | null = memoizedState.next;
+    if (!nextState) break;
+    memoizedState = nextState;
+  }
+
+  return false;
+};
+
+export const isPromise = (value: unknown): value is Promise<unknown> => {
+  if (!value || typeof value !== 'object') {
+    return false;
+  }
+
+  try {
+    return (
+      value instanceof Promise ||
+      ('then' in value &&
+        typeof (value as { then: unknown }).then === 'function')
+    );
+  } catch {
+    return false;
+  }
+};
+
+export const ensureRecord = (
+  value: unknown,
+  maxDepth = 2,
+  seen = new WeakSet<object>(),
+): Record<string, unknown> => {
+  if (isPromise(value)) {
+    return { type: 'promise', displayValue: 'Promise' };
+  }
+
+  if (value === null) {
+    return { type: 'null', displayValue: 'null' };
+  }
+
+  if (value === undefined) {
+    return { type: 'undefined', displayValue: 'undefined' };
+  }
+
+  switch (typeof value) {
+    case 'object': {
+      if (seen.has(value)) {
+        return { type: 'circular', displayValue: '[Circular Reference]' };
+      }
+
+      if (!value) return { type: 'null', displayValue: 'null' };
+
+      seen.add(value);
+
+      try {
+        const result: Record<string, unknown> = {};
+
+        if (value instanceof Element) {
+          result.type = 'Element';
+          result.tagName = value.tagName.toLowerCase();
+          result.displayValue = value.tagName.toLowerCase();
+          return result;
+        }
+
+        if (value instanceof Map) {
+          result.type = 'Map';
+          result.size = value.size;
+          result.displayValue = `Map(${value.size})`;
+
+          if (maxDepth > 0) {
+            const entries: Record<string, unknown> = {};
+            let index = 0;
+            for (const [key, val] of value.entries()) {
+              if (index >= 50) break;
+              try {
+                entries[String(key)] = ensureRecord(val, maxDepth - 1, seen);
+              } catch {
+                entries[String(index)] = {
+                  type: 'error',
+                  displayValue: 'Error accessing Map entry',
+                };
+              }
+              index++;
+            }
+            result.entries = entries;
+          }
+          return result;
+        }
+
+        if (value instanceof Set) {
+          result.type = 'Set';
+          result.size = value.size;
+          result.displayValue = `Set(${value.size})`;
+
+          if (maxDepth > 0) {
+            const items = [];
+            let count = 0;
+            for (const item of value) {
+              if (count >= 50) break;
+              items.push(ensureRecord(item, maxDepth - 1, seen));
+              count++;
+            }
+            result.items = items;
+          }
+          return result;
+        }
+
+        if (value instanceof Date) {
+          result.type = 'Date';
+          result.value = value.toISOString();
+          result.displayValue = value.toLocaleString();
+          return result;
+        }
+
+        if (value instanceof RegExp) {
+          result.type = 'RegExp';
+          result.value = value.toString();
+          result.displayValue = value.toString();
+          return result;
+        }
+
+        if (value instanceof Error) {
+          result.type = 'Error';
+          result.name = value.name;
+          result.message = value.message;
+          result.displayValue = `${value.name}: ${value.message}`;
+          return result;
+        }
+
+        if (value instanceof ArrayBuffer) {
+          result.type = 'ArrayBuffer';
+          result.byteLength = value.byteLength;
+          result.displayValue = `ArrayBuffer(${value.byteLength})`;
+          return result;
+        }
+
+        if (value instanceof DataView) {
+          result.type = 'DataView';
+          result.byteLength = value.byteLength;
+          result.displayValue = `DataView(${value.byteLength})`;
+          return result;
+        }
+
+        if (ArrayBuffer.isView(value)) {
+          const typedArray = value as unknown as {
+            length: number;
+            constructor: { name: string };
+            buffer: ArrayBuffer;
+          };
+          result.type = typedArray.constructor.name;
+          result.length = typedArray.length;
+          result.byteLength = typedArray.buffer.byteLength;
+          result.displayValue = `${typedArray.constructor.name}(${typedArray.length})`;
+          return result;
+        }
+
+        if (Array.isArray(value)) {
+          result.type = 'array';
+          result.length = value.length;
+          result.displayValue = `Array(${value.length})`;
+
+          if (maxDepth > 0) {
+            result.items = value
+              .slice(0, 50)
+              .map((item) => ensureRecord(item, maxDepth - 1, seen));
+          }
+          return result;
+        }
+
+        const keys = Object.keys(value);
+        result.type = 'object';
+        result.size = keys.length;
+        result.displayValue =
+          keys.length <= 5
+            ? `{${keys.join(', ')}}`
+            : `{${keys.slice(0, 5).join(', ')}, ...${keys.length - 5}}`;
+
+        if (maxDepth > 0) {
+          const entries: Record<string, unknown> = {};
+          for (const key of keys.slice(0, 50)) {
+            try {
+              entries[key] = ensureRecord(
+                (value as Record<string, unknown>)[key],
+                maxDepth - 1,
+                seen,
+              );
+            } catch {
+              entries[key] = {
+                type: 'error',
+                displayValue: 'Error accessing property',
+              };
+            }
+          }
+          result.entries = entries;
+        }
+        return result;
+      } finally {
+        seen.delete(value);
+      }
+    }
+    case 'string':
+      return {
+        type: 'string',
+        value,
+        displayValue: `"${value}"`,
+      };
+    case 'function':
+      return {
+        type: 'function',
+        displayValue: 'ƒ()',
+        name: value.name || 'anonymous',
+      };
+    default:
+      return {
+        type: typeof value,
+        value,
+        displayValue: String(value),
+      };
+  }
+};
+
+export const getCurrentFiberState = (
+  fiber: Fiber,
+): Record<string, unknown> | null => {
+  if (fiber.tag !== FunctionComponentTag || !isDirectComponent(fiber)) {
+    return null;
+  }
+
+  const currentIsNewer = fiber.alternate
+    ? (fiber.actualStartTime ?? 0) > (fiber.alternate.actualStartTime ?? 0)
+    : true;
+
+  const memoizedState: ExtendedMemoizedState | null = currentIsNewer
+    ? fiber.memoizedState
+    : (fiber.alternate?.memoizedState ?? fiber.memoizedState);
+
+  if (!memoizedState) return null;
+
+  return memoizedState;
 };
 
 export const replayComponent = async (fiber: Fiber): Promise<void> => {
